rule_generation/requirements.txt
pypdf
langchain
langchain_community
# os
langchain_groq
python-dotenv
langchain_text_splitters
streamlit
